Correct thread iteration order
The behavior of this fix can be verified by running:
  cargo run -- examples/spawn-loop.poline
and verifying that an infinite loop occurrs.

The intended iteration order is to start with the current thread,
and then iterate forward, wrapping around if necessary, back to this
starting thread. The new current thread is whatever thread iteration
stops at.

Previously, this order was not respected, which meant that programs
which should have gotten into infinite loops simply stopped instead.

With this iteration order, and replacing threads that have finished
running with ThreadState::Done, we can keep a constant number of threads
when constantly spawning, as long as threads finish quickly.