Prevent threads from reusing identifiers
Previously threads were identified based on what slot they happened
to occupy in our vector of threads. This was problematic because
it could lead to a message sent to a dead thread triggering a live
thread that had replaced its slot. We want to replace slots
to not use up memory for threads with no more code to run, and we
want to avoid sending to a dead thread to trigger anything.

This separates logical index from the slot a thread happens to occupy.
Logical indices are always increasing, and are unique.

One potential downside is that we're limited to spawning 2^32 threads,
but we can switch to using u64 if we encounter that limit, which is unlikely.